null: Use the same code for both vfs versions
